+++ extension/httpfs/src/include/http_config.h
@@ -12,8 +12,8 @@ struct HTTPConfig {
 };
 
 struct HTTPCacheInMemoryConfig {
-    static constexpr const char* HTTP_CACHE_IN_MEMORY_ENV_VAR = "HTTP_CACHE_IN_MEMORY";
-    static constexpr const char* HTTP_CACHE_IN_MEMORY_OPTION = "http_cache_in_memory";
+    static constexpr const char* HTTP_CACHE_FILE_ENV_VAR = "HTTP_CACHE_FILE";
+    static constexpr const char* HTTP_CACHE_FILE_OPTION = "http_cache_file";
     static constexpr bool DEFAULT_CACHE_IN_MEMORY = false;
 };
