+++ extension/httpfs/src/include/s3fs.h
@@ -50,7 +50,7 @@ struct S3FileInfo final : public HTTPFileInfo {
 
     ~S3FileInfo() override;
 
-    void initialize(main::ClientContext* context) override;
+    void initialize() override;
 
     void initializeClient() override;
